Add column printing scripts from Section 3.4
These scripts (one script all linked to the same file) have been
modified from the text. The scripts directly from the text as-is don't
work. The reason is probably because the test system does not include
the current directory (.) by default in the executable search path.
Adding the path to these scripts into the PATH variable breaks the `pr`
call. The modifications from the text fixes the issue.