Add word count app from K&R2 Chapter 6
For exercise 4-4: comparing a C word count implementation with the shell
pipeline implementation. This version is almost straight from the text
and doesn't have the top 10 list like the shell version. The top 10
list version has already been started but putting this version here
for reference since this is the starting point.