Solution 4-10: Incomplete solution for part 1
This is a convenience check in for part one of the solution.
This code splits lines on word boundaries but has some issues. It
doesn't justify the last chunk of a line to the correct position.
It doesn't deal with words longer than a line.
It's sort of confusing and probably not the best approach.