calendar: Fix date format to match text
This version of calendar is finally working. No need to send mail
because there's no terminal mail client set up on the test system.
The format of the `date` command had to change from the test system's
default to match what the text expects.
Run as:
$ ./calendar.1
Make sure the calendar file is in $HOME.