Solution 5-15: Add diff for keeping builtin source
This diff probably cannot be directly applied by
patch. This is simply to show how one might edit
the Bash source builtins Makefile such that intermediate
C source files that implement the builtin functions
are maintained. This was done for exercise 5-15 in order
to example how the Bash shell implements the times builtin
command.